news/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DAWG-Python==0.7.2
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
feedparser==6.0.11
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.5.4
gensim==4.3.2
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
intervaltree==3.1.0
ipymarkup==0.9.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
Keras-Preprocessing==1.1.2
libclang==18.1.1
lxml==5.2.1
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.5
namex==0.0.8
natasha==1.6.0
navec==0.10.0
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
packaging==24.0
pandas==2.2.2
passlib==1.7.4
pika==1.3.2
protobuf==4.25.3
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
razdel==0.5.0
regex==2024.5.10
requests==2.31.0
rich==13.7.1
scikit-learn==1.2.2
scipy==1.10.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
slovnet==0.6.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yargy==0.16.0
yarl==1.9.4